attempted to change spent value from int to float to allow decimals but am getting error, also added some back to readme

+++ server/utils/playground.js
@@ -54,7 +54,7 @@ query User($id: ID!) {
     }
   }
   
-  mutation ADD_RECEIPT($user: String!,$spent: Int, $purchaseDate: String!, $place: String!) {
+  mutation ADD_RECEIPT($user: String!,$spent: Float, $purchaseDate: String!, $place: String!) {
     addReceipt(user: $user, spent: $spent, purchaseDate: $purchaseDate, place: $place) {
       _id
       spent
